Roombound research
Found more accessors for cell nodes.

Found TESObjectREFR::SetExtraMultiBoundRef. Code near this appears to control what roombound a reference renders under. I can't find exactly where the game runs bounds checks against the reference's position, but I've decoded enough of BSMultiBoundShape to just walk a cell's rooms and run those checks myself.

BSMultiBoundShape and SOME of its subclasses have been documented. At least one (BSMultiBoundCapsule) remains undocumented.

+++ ReverseEngineered/NetImmerse/nodes.h
@@ -33,7 +33,7 @@ namespace RE {
       public:
          NiPointer<BSMultiBound> unkB8 = nullptr;
          uint32_t unkBC = 0;
-         uint32_t unkC0 = 0;
+         float    unkC0 = 0.0F;
          //
          MEMBER_FN_PREFIX(BSMultiBoundNode);
          DEFINE_MEMBER_FN(Constructor, BSMultiBoundNode&, 0x00B16110);
